refactor: comment reviews, spacing and removed alert maxWidth
