+++ apps/main/src/llamalend/features/market-position-details/HealthDetails.tsx
@@ -3,7 +3,7 @@ import { Metric } from '@ui-kit/shared/ui/Metric'
 import { Tooltip } from '@ui-kit/shared/ui/Tooltip'
 import { SizesAndSpaces } from '@ui-kit/themes/design/1_sizes_spaces'
 import { decimal } from '@ui-kit/utils'
-import { HEALTH_TOOLTIP } from './tooltips/health-tooltip'
+import { HEALTH_TOOLTIP } from './tooltips'
 import { HealthBar, getHealthValueColor, Health, LiquidationAlert } from './'
 
 const { Spacing } = SizesAndSpaces
@@ -16,7 +16,7 @@ export const HealthDetails = ({
   liquidationAlert: LiquidationAlert
 }) => {
   const theme = useTheme()
-  const { title, body, placement } = HEALTH_TOOLTIP
+  const { title, body } = HEALTH_TOOLTIP
 
   return (
     <Stack>
@@ -24,15 +24,15 @@ export const HealthDetails = ({
         <Stack direction="row" alignItems="flex-end" gap={Spacing.md.mobile}>
           <Metric
             label={title}
-            value={value != null ? Number(value) : null}
+            value={value}
             loading={loading}
             valueOptions={{ unit: 'none', color: getHealthValueColor({ health: decimal(value), theme }) }}
             valueTooltip={HEALTH_TOOLTIP}
             size="medium"
           />
-          <Tooltip title={title} body={body} placement={placement}>
+          <Tooltip title={title} body={body}>
             <Stack sx={{ flex: 1 }}>
-              <HealthBar health={value != null ? Number(value) : null} softLiquidation={softLiquidation} />
+              <HealthBar health={value} softLiquidation={softLiquidation} />
             </Stack>
           </Tooltip>
         </Stack>
