+++ apps/main/src/llamalend/features/market-position-details/tooltips/index.ts
@@ -1,4 +1,13 @@
+import { createElement } from 'react'
+import { t } from '@ui-kit/lib/i18n'
+import { HealthTooltipContent } from './HealthTooltipContent'
+
 export * from './AmountSuppliedTooltipContent'
-export * from './health-tooltip'
+export * from './HealthTooltipContent'
 export * from './LiquidationThresholdMetricTooltipContent'
 export * from './VaultSharesTooltipContent'
+
+export const HEALTH_TOOLTIP = {
+  title: t`Health`,
+  body: createElement(HealthTooltipContent),
+} as const
