+++ packages/curve-ui-kit/src/themes/components/mui-alert.tsx
@@ -35,7 +35,6 @@ export const defineMuiAlert = (
       }),
       '& .MuiAlert-message': {
         flexGrow: 1,
-        maxWidth: '90ch',
         textWrapStyle: 'pretty',
         ...handleBreakpoints({
           paddingInlineEnd: Spacing.sm,
